+++ src/main/java/com/jj/swm/domain/study/repository/CustomStudyBookmarkRepositoryImpl.java
@@ -10,7 +10,7 @@ import java.util.UUID;
 
 import static com.jj.swm.domain.study.entity.QStudy.study;
 import static com.jj.swm.domain.study.entity.QStudyBookmark.studyBookmark;
-import static com.jj.swm.domain.user.entity.QUser.user;
+import static com.jj.swm.domain.user.core.entity.QUser.user;
 
 @RequiredArgsConstructor
 public class CustomStudyBookmarkRepositoryImpl implements CustomStudyBookmarkRepository {
