+++ src/main/java/com/jj/swm/domain/studyroom/core/repository/custom/impl/CustomStudyRoomBookmarkRepositoryImpl.java
@@ -11,7 +11,7 @@ import java.util.UUID;
 
 import static com.jj.swm.domain.studyroom.core.entity.QStudyRoom.studyRoom;
 import static com.jj.swm.domain.studyroom.core.entity.QStudyRoomBookmark.studyRoomBookmark;
-import static com.jj.swm.domain.user.entity.QUser.user;
+import static com.jj.swm.domain.user.core.entity.QUser.user;
 
 @RequiredArgsConstructor
 public class CustomStudyRoomBookmarkRepositoryImpl implements CustomStudyRoomBookmarkRepository {
